backend/requirements.txt
fastapi==0.104.1
langchain==0.1.0
langchain-google-genai==0.0.10
pydantic==2.5.0
uvicorn==0.24.0
hypothesis==6.92.0
pytest==7.4.3
python-dotenv==1.0.0
